requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.17.0
python-dotenv>=1.0.1
mkdocs>=1.6.0
mkdocs-material>=9.5.0
